+++ src/screenshot-diff/screenshot-diff-overlay-matches.ts
@@ -2,8 +2,8 @@ import type {
   ScreenshotDiffRegion,
   ScreenshotDiffRegionOverlayMatch,
 } from './screenshot-diff-regions.ts';
-import type { Rect, ScreenshotOverlayRef } from './snapshot.ts';
-import { intersectArea } from './screenshot-geometry.ts';
+import type { Rect, ScreenshotOverlayRef } from '../utils/snapshot.ts';
+import { intersectArea } from '../utils/screenshot-geometry.ts';
 
 const MAX_MATCHES_PER_REGION = 3;
