+++ src/screenshot-diff/screenshot-diff-region-split.ts
@@ -1,4 +1,4 @@
-import type { PNG } from './png.ts';
+import type { PNG } from '../utils/png.ts';
 import type { MutableDiffRegion } from './screenshot-diff-region-types.ts';
 
 // Region splitting is based on screen-relative heights so it works on phone,
