+++ src/screenshot-diff/screenshot-diff-region-types.ts
